Rename file to events.js
